refactor(server/schemas): Moved schemas into separate modules
Moved so that queries and mutations are separated from all of the base
definitions. This makes it cleaner to read the code and find specific
definitions
